+++ native/src/main/java/io/ballerina/stdlib/http/api/HttpDispatcher.java
@@ -91,7 +91,8 @@ public class HttpDispatcher {
             PredefinedTypes.TYPE_JSON));
     private static final MapType STRING_MAP = TypeCreator.createMapType(PredefinedTypes.TYPE_STRING);
 
-    public static HttpService findService(HTTPServicesRegistry servicesRegistry, HttpCarbonMessage inboundReqMsg) {
+    public static HttpService findService(HTTPServicesRegistry servicesRegistry, HttpCarbonMessage inboundReqMsg,
+                                          boolean forInterceptors) {
         try {
             Map<String, HttpService> servicesOnInterface;
             List<String> sortedServiceURIs;
@@ -110,12 +111,8 @@ public class HttpDispatcher {
             }
 
             String rawUri = (String) inboundReqMsg.getProperty(HttpConstants.TO);
-            inboundReqMsg.setProperty(HttpConstants.RAW_URI, rawUri);
             Map<String, Map<String, String>> matrixParams = new HashMap<>();
             String uriWithoutMatrixParams = URIUtil.extractMatrixParams(rawUri, matrixParams);
-
-            inboundReqMsg.setProperty(HttpConstants.TO, uriWithoutMatrixParams);
-            inboundReqMsg.setProperty(HttpConstants.MATRIX_PARAMS, matrixParams);
 
             URI validatedUri = getValidatedURI(uriWithoutMatrixParams);
 
@@ -129,7 +126,12 @@ public class HttpDispatcher {
             }
 
             HttpService service = servicesOnInterface.get(basePath);
-            setInboundReqProperties(inboundReqMsg, validatedUri, basePath);
+            if (!forInterceptors) {
+                setInboundReqProperties(inboundReqMsg, validatedUri, basePath);
+                inboundReqMsg.setProperty(HttpConstants.RAW_URI, rawUri);
+                inboundReqMsg.setProperty(HttpConstants.TO, uriWithoutMatrixParams);
+                inboundReqMsg.setProperty(HttpConstants.MATRIX_PARAMS, matrixParams);
+            }
             return service;
         } catch (Exception e) {
             throw new BallerinaConnectorException(e.getMessage());
@@ -216,7 +218,7 @@ public class HttpDispatcher {
 
         try {
             // Find the Service TODO can be improved
-            HttpService service = HttpDispatcher.findService(servicesRegistry, inboundMessage);
+            HttpService service = HttpDispatcher.findService(servicesRegistry, inboundMessage, false);
             if (service == null) {
                 throw new BallerinaConnectorException("no Service found to handle the service request");
                 // Finer details of the errors are thrown from the dispatcher itself, Ideally we shouldn't get here.
@@ -291,6 +293,10 @@ public class HttpDispatcher {
                     } else {
                         requestCtx.addNativeData(HttpConstants.INTERCEPTOR_SERVICE, false);
                     }
+                    int interceptorId = httpCarbonMessage.getProperty(HttpConstants.INTERCEPTOR_SERVICE_INDEX) == null
+                            ? 0 : (int) httpCarbonMessage.getProperty(HttpConstants.INTERCEPTOR_SERVICE_INDEX) - 1;
+                    requestCtx.addNativeData(HttpConstants.INTERCEPTOR_SERVICE_INDEX, interceptorId);
+                    requestCtx.addNativeData(HttpConstants.REQUEST_CONTEXT_NEXT, false);
                     index = ((NonRecurringParam) param).getIndex();
                     paramFeed[index++] = requestCtx;
                     paramFeed[index] = true;
@@ -482,8 +488,9 @@ public class HttpDispatcher {
 
     static BObject getCaller(Resource resource, HttpCarbonMessage httpCarbonMessage,
                                 BMap<BString, Object> endpointConfig) {
-        BObject httpCaller = httpCarbonMessage.getProperty(HttpConstants.CALLER) == null ?
-                ValueCreatorUtils.createCallerObject() : (BObject) httpCarbonMessage.getProperty(HttpConstants.CALLER);
+        final BObject httpCaller = httpCarbonMessage.getProperty(HttpConstants.CALLER) == null ?
+                ValueCreatorUtils.createCallerObject(httpCarbonMessage) :
+                (BObject) httpCarbonMessage.getProperty(HttpConstants.CALLER);
         HttpUtil.enrichHttpCallerWithConnectionInfo(httpCaller, httpCarbonMessage, resource, endpointConfig);
         HttpUtil.enrichHttpCallerWithNativeData(httpCaller, httpCarbonMessage, endpointConfig);
         httpCarbonMessage.setProperty(HttpConstants.CALLER, httpCaller);
@@ -492,14 +499,12 @@ public class HttpDispatcher {
 
     static BObject createRequestContext(HttpCarbonMessage httpCarbonMessage, BMap<BString, Object> endpointConfig) {
         BObject requestContext = ValueCreatorUtils.createRequestContextObject();
-        BArray interceptors = endpointConfig.getArrayValue(HttpConstants.ENDPOINT_CONFIG_INTERCEPTORS);
-        if (interceptors != null) {
-            requestContext.addNativeData(HttpConstants.HTTP_INTERCEPTORS, interceptors);
-        }
+        BArray interceptors = httpCarbonMessage.getProperty(HttpConstants.INTERCEPTORS) instanceof BArray ?
+                              (BArray) httpCarbonMessage.getProperty(HttpConstants.INTERCEPTORS) : null;
+        requestContext.addNativeData(HttpConstants.INTERCEPTORS, interceptors);
+        requestContext.addNativeData(HttpConstants.TARGET_SERVICE, httpCarbonMessage.getProperty(
+                                     HttpConstants.TARGET_SERVICE));
         requestContext.addNativeData(HttpConstants.REQUEST_CONTEXT_NEXT, false);
-        int interceptorId = httpCarbonMessage.getProperty(HttpConstants.INTERCEPTOR_SERVICE_INDEX) == null
-                ? 0 : (int) httpCarbonMessage.getProperty(HttpConstants.INTERCEPTOR_SERVICE_INDEX) - 1;
-        requestContext.addNativeData(HttpConstants.INTERCEPTOR_SERVICE_INDEX, interceptorId);
         httpCarbonMessage.setProperty(HttpConstants.REQUEST_CONTEXT, requestContext);
         return requestContext;
     }