+++ native/src/main/java/io/ballerina/stdlib/http/api/HttpInterceptorUnitCallback.java
@@ -18,6 +18,7 @@
 
 package io.ballerina.stdlib.http.api;
 
+import io.ballerina.runtime.api.PredefinedTypes;
 import io.ballerina.runtime.api.Runtime;
 import io.ballerina.runtime.api.async.Callback;
 import io.ballerina.runtime.api.values.BArray;
@@ -25,8 +26,6 @@ import io.ballerina.runtime.api.values.BError;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.stdlib.http.api.nativeimpl.ModuleUtils;
 import io.ballerina.stdlib.http.transport.message.HttpCarbonMessage;
-
-import static java.lang.System.err;
 
 /**
  * {@code HttpInterceptorUnitCallback} is the responsible for acting on notifications received from Ballerina side when
@@ -77,7 +76,7 @@ public class HttpInterceptorUnitCallback implements Callback {
 
     public void sendFailureResponse(BError error) {
         cleanupRequestAndContext();
-        HttpUtil.handleFailure(requestMessage, error);
+        HttpUtil.handleFailure(requestMessage, error, false);
     }
 
     private void cleanupRequestAndContext() {
@@ -101,13 +100,12 @@ public class HttpInterceptorUnitCallback implements Callback {
 
     private void sendRequestToNextService() {
         ballerinaHTTPConnectorListener.onMessage(requestMessage);
-        requestCtx.addNativeData(HttpConstants.REQUEST_CONTEXT_NEXT, false);
     }
 
     private void validateResponseAndProceed(Object result) {
         int interceptorId = (int) requestCtx.getNativeData(HttpConstants.INTERCEPTOR_SERVICE_INDEX);
         requestMessage.setProperty(HttpConstants.INTERCEPTOR_SERVICE_INDEX, interceptorId);
-        BArray interceptors = (BArray) requestCtx.getNativeData(HttpConstants.HTTP_INTERCEPTORS);
+        BArray interceptors = (BArray) requestCtx.getNativeData(HttpConstants.INTERCEPTORS);
         boolean nextCalled = (boolean) requestCtx.getNativeData(HttpConstants.REQUEST_CONTEXT_NEXT);
 
         if (alreadyResponded(result)) {
@@ -136,6 +134,15 @@ public class HttpInterceptorUnitCallback implements Callback {
                             "with the configuration", HttpErrorType.GENERIC_LISTENER_ERROR);
                     sendFailureResponse(err);
                 }
+            } else {
+                Object targetService = requestCtx.getNativeData(HttpConstants.TARGET_SERVICE);
+                if (result.equals(targetService)) {
+                    sendRequestToNextService();
+                } else {
+                    BError err = HttpUtil.createHttpError("target service did not match with the configuration",
+                            HttpErrorType.GENERIC_LISTENER_ERROR);
+                    sendFailureResponse(err);
+                }
             }
         }
     }
@@ -160,7 +167,8 @@ public class HttpInterceptorUnitCallback implements Callback {
                 sendFailureResponse(result);
             }
         };
-        runtime.invokeMethodAsync(caller, "returnResponse", null,
-                ModuleUtils.getNotifySuccessMetaData(), returnCallback, paramFeed);
+        runtime.invokeMethodAsyncSequentially(
+                caller, "returnResponse", null, ModuleUtils.getNotifySuccessMetaData(),
+                returnCallback, null, PredefinedTypes.TYPE_NULL, paramFeed);
     }
 }